% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read_xvg}
\alias{read_xvg}
\title{Read xvg file}
\usage{
read_xvg(str)
}
\arguments{
\item{str}{:: filepath to the xvg file}
}
\value{
cleaned data  :: tibble
}
\description{
And also cleanup and parse data
}
